# microservices/api-gateway/openapi.yml
openapi: 3.0.3
info:
  title: POS Multi-Magasins - API Gateway
  description: |
    API Gateway pour le système POS Multi-Magasins avec architecture microservices.
    
    ## Authentification
    
    ### API Keys (Kong Gateway)
    - **Header**: `X-API-KEY`
    - **Valeur**: `pos-microservices-2025`
    
    ### JWT Authentication (Services clients)
    - **Header**: `Authorization`
    - **Format**: `Bearer <jwt_token>`
    - **Endpoint de login**: `POST /api/v1/customers/auth/login`
    
    ## Architecture Microservices
    
    Le système est composé de 7 microservices :
    
    ### Services Métier (extraits du monolithe)
    1. **Product Service** (Port 8001) - Catalogue produits
    2. **Inventory Service** (Port 8002) - Gestion des stocks  
    3. **Sales Service** (Port 8003) - Transactions POS
    4. **Reporting Service** (Port 8004) - Rapports analytiques
    
    ### Services E-commerce (nouveaux)
    5. **Customer Service** (Port 8005) - Comptes clients
    6. **Cart Service** (Port 8006) - Panier d'achats
    7. **Order Service** (Port 8007) - Gestion des commandes
    
    ## Load Balancing
    
    Le Cart Service utilise un load balancer avec 3 instances (algorithme least-connections).
    
    ## Monitoring
    
    - **Prometheus**: Métriques sur `/metrics`
    - **Grafana**: Dashboard Kong Gateway
    - **Health checks**: `/health` sur chaque service
    
  version: "1.0.0"
  contact:
    name: LOG430 Lab 5 Team
    email: lab430@example.com
  license:
...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: Clé API pour accéder aux microservices via Kong Gateway
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT pour l'authentification des clients
  schemas:
...
    CartItem:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
          description: ID du produit
        quantity:
          type: integer
          minimum: 1
          description: Quantité
        price:
          type: number
          format: float
          description: Prix unitaire
        name:
          type: string
          description: Nom du produit
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        total:
          type: number
          format: float
          description: Total du panier
        customer_id:
          type: integer
          description: ID du client propriétaire
    
    # Modèles Order Service
...
paths:
...
  /api/v1/cart:
    get:
      tags: [Cart]
      summary: Récupérer le panier
      description: Obtenir le contenu du panier (load balancé sur 3 instances)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
          description: ID de session du panier
      responses:
        '200':
          description: Contenu du panier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
    
    post:
      tags: [Cart]
      summary: Ajouter un produit au panier
      description: Ajouter ou mettre à jour un produit dans le panier
      security:
        - ApiKeyAuth: []
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id, quantity]
              properties:
                product_id:
                  type: integer
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Produit ajouté au panier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'

    put:
      tags: [Cart]
      summary: Mettre à jour la quantité
      description: Modifier la quantité d'un produit dans le panier
      security:
        - ApiKeyAuth: []
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id, quantity]
              properties:
                product_id:
                  type: integer
                quantity:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Quantité mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
...
  /api/v1/cart/{customer_id}:
    get:
      tags: [Cart]
      summary: Récupérer le panier
      description: Obtenir le contenu du panier d'un client (Load Balanced - 3 instances)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Contenu du panier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
    
    post:
      tags: [Cart]
      summary: Ajouter au panier
      description: Ajouter un produit au panier (Load Balanced - 3 instances)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItem'
      responses:
        '200':
          description: Produit ajouté au panier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'

  # Order Service Routes
...